add win32.h file
